Add nonce reuse to the constant-value detection end-to-end test
Extend the RFC 9700 Section 2.1.1 end-to-end coverage to the OpenID Connect
nonce, symmetric to the code_challenge case: reusing a nonce (with a fresh
code_challenge on each request) across two authorizations is rejected with
invalid_request. Both cases now live in one class,
AuthorizationValueReuseDetectionE2ETests, sharing the isolated-host and
authorize-parameter helpers.
